requirements.txt
mediapipe==0.10.9
opencv-python==4.9.0.80
numpy==1.26.4
scipy==1.13.0
PyQt5==5.15.10
paho-mqtt==2.0.0
plyer==2.1.0
matplotlib==3.8.4
ultralytics==8.2.0
# sqlite3 is built into Python — no install needed
reportlab==4.2.0
Flask==3.0.3
flask-socketio==5.3.6
eventlet==0.35.2
gunicorn==21.2.0
psutil==5.9.8
